requirements.txt
asgiref==3.7.2
Django==5.0.1
django-appconf==1.0.6
django-browser-reload==1.12.1
django-ckeditor==6.7.0
django-compressor==4.4
django-js-asset==2.2.0
django-modeltranslation==0.18.11
django-tinymce==3.6.1
pillow==10.2.0
rcssmin==1.1.1
rjsmin==1.2.1
six==1.16.0
sqlparse==0.4.4
tinycss2==1.1.1
typing_extensions==4.9.0
tzdata==2023.4
webencodings==0.5.1
